[posts] check if user has modification rights on the post

+++ src/posts/posts.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Post } from 'src/db/entities/post.entity';
 import { Repository } from 'typeorm';
@@ -22,25 +26,41 @@ export class PostsService {
     return this.postsRepository.save(newPost);
   }
 
-  async deletePost(id: number): Promise<Object> {
-    try {
-      const post = await this.postsRepository.findOneOrFail(id);
-      this.postsRepository.remove(post);
-      return {
-        statusCode: 200,
-        message: 'Post deleted',
-      };
-    } catch {
-      throw new NotFoundException(`Post with id ${id} does not exist!`);
+  async deletePost(postId: number, userId: number): Promise<Object> {
+    const post = await this.postsRepository.findOne(postId);
+
+    if (!post) {
+      throw new NotFoundException(`Post with id ${postId} does not exist!`);
     }
+
+    if (post.authorId !== userId) {
+      throw new UnauthorizedException(
+        `User with id ${userId} has no modification rights on this post`,
+      );
+    }
+
+    this.postsRepository.remove(post);
+
+    return { statusCode: 200, message: 'Post deleted' };
   }
 
-  async updatePost(id: number, payload: PostPayload): Promise<PostResponse> {
-    try {
-      const post = await this.postsRepository.findOneOrFail(id);
-      return this.postsRepository.save({ ...post, ...payload });
-    } catch {
-      throw new NotFoundException(`Post with id ${id} does not exist!`);
+  async updatePost(
+    postId: number,
+    userId: number,
+    payload: PostPayload,
+  ): Promise<PostResponse> {
+    const post = await this.postsRepository.findOne(postId);
+
+    if (!post) {
+      throw new NotFoundException(`Post with id ${postId} does not exist!`);
     }
+
+    if (post.authorId !== userId) {
+      throw new UnauthorizedException(
+        `User with id ${userId} has no modification rights on this post`,
+      );
+    }
+
+    return this.postsRepository.save({ ...post, ...payload });
   }
 }
